requirements.txt
# This file was generated by mk-requires.sh

ansi2html
argparse
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure==4.0.0
backports-abc
bandit
boto3
botocore
certifi
cffi
chardet
croniter
cryptography
daemon
futures
gitdb2
gitpython
gprof2dot
idna
jinja2
markupsafe
mock
msgpack
ntplib
objgraph
patch==1.*
pycryptodome
pygit2<0.27.0
pyinotify
pyinstaller-hooks
pyinstaller==3.6
pylint
pyopenssl>=16.2.0
pytest
pytest-cov
pytest-html
pytest-pythonpath
python-dateutil
pyyaml
requests>=2.13.0
salt-ssh==2019.2.3
saltpylint
singledispatch
six
smmap2
tornado
urllib3
vulners
